Revert "Accept an integer"
This reverts commit 18d2cce425962cc1951752409c16f54fba41c2b2.

+++ homeassistant/components/thermostat/__init__.py
@@ -52,7 +52,7 @@ SET_AWAY_MODE_SCHEMA = vol.Schema({
 })
 SET_TEMPERATURE_SCHEMA = vol.Schema({
     vol.Optional(ATTR_ENTITY_ID): cv.entity_ids,
-    vol.Required(ATTR_TEMPERATURE): vol.Coerce(int),
+    vol.Required(ATTR_TEMPERATURE): vol.Coerce(float),
 })
 SET_FAN_MODE_SCHEMA = vol.Schema({
     vol.Optional(ATTR_ENTITY_ID): cv.entity_ids,
